+++ 001-100/001-010/Problem010/Problem010.cpp
@@ -17,7 +17,7 @@
 
 int main()
 {
-	std::vector<unsigned> primes = primeEratosthenes( 2000000 );
+	std::vector<uint64_t> primes = primeEratosthenes( 2000000 );
 
 	std::cout << "Answer: " << std::accumulate( primes.begin(), primes.end(), uint64_t( 0 ), []( uint64_t a, uint64_t b ) -> uint64_t { return a + b; } ) << std::endl;
 	std::cin.get();
